feat!: major update Stories and StoriesGroup
