+++ src/components/Stories/README.md
@@ -14,25 +14,37 @@ Component for displaying stories. It looks like a carousel in a modal with given
 | onNextClick         | `Function`      |          |         | Action when switching to next story              |
 | disableOutsideClick | `Boolean`       |          | true    | If `true`, do not close stories on click outside |
 | className           | `string`        |          |         | Stories modal class                              |
-| action              | `ButtonProps`   |          |         | Custom action button props for the last step     |
 
 ### StoriesItem object
 
-| Field       | Type               | Required | Default | Description                      |
-| ----------- | ------------------ | -------- | ------- | -------------------------------- |
-| title       | `String`           |          |         | Title                            |
-| description | `String`           |          |         | Main text, deprecated            |
-| content     | `React.ReactNode`  |          |         | Main content                     |
-| url         | `String`           |          |         | Link to display more information |
-| media       | `StoriesItemMedia` |          |         | Media content                    |
+| Field           | Type                     | Required | Default | Description                              |
+| --------------- | ------------------------ | -------- | ------- | ---------------------------------------- |
+| title           | `String`                 |          |         | Title                                    |
+| description     | `String`                 |          |         | Main text                                |
+| url             | `String`                 |          |         | Link to display more information         |
+| media           | `StoriesItemMedia`       |          |         | Media content                            |
+| firstAction     | `ButtonProps`            |          |         | Custom action button props               |
+| secondAction    | `ButtonProps`            |          |         | Custom action button props               |
+| textBlockStyle  | `StoriesTextBlockStyle`  | ✓        |         | Props for styling text content in Story  |
+| mediaBlockStyle | `StoriesMediaBlockStyle` | ✓        |         | Props for styling media content in Story |
+| textColorStyles | `StoriesItemTextStyles`  |          |         | Props for styling text color in Story    |
 
 ### StoriesItemMedia object
 
-| Field     | Type     | Required | Default | Description                       |
-| --------- | -------- | -------- | ------- | --------------------------------- |
-| type      | `String` |          | image   | Content type (`image` or `video`) |
-| url       | `String` | ✓        |         | File link                         |
-| posterUrl | `String` |          |         | Poster URL (only used for video)  |
+| Field     | Type     | Required | Default | Description                                        |
+| --------- | -------- | -------- | ------- | -------------------------------------------------- |
+| type      | `String` | ✓        |         | Content type (`image` or `video`)                  |
+| url       | `String` | ✓        |         | File link                                          |
+| url2x     | `String` |          |         | File link for Retina display (only used for image) |
+| posterUrl | `String` |          |         | Poster URL (only used for video)                   |
+
+### StoriesItemTextStyles object
+
+| Field            | Type     | Required | Default | Description                      |
+| ---------------- | -------- | -------- | ------- | -------------------------------- |
+| titleColor       | `String` |          |         | Apply color to Story title       |
+| descriptionColor | `String` |          |         | Apply color to Story description |
+| counterColor     | `String` |          |         | Apply color to Story counter     |
 
 #### Usage example
 
@@ -42,7 +54,8 @@ Component for displaying stories. It looks like a carousel in a modal with given
   items={[
     {
       title: 'Story title',
-      content: <b>Story text</b>,
+      description: 'Story text',
+      type: 'image',
       media: {
         url: 'https://storage.yandexcloud.net/uikit-storybook-assets/story-picture-2.png',
       },
