+++ src/components/Stories/components/ImageView/ImageView.tsx
@@ -1,16 +1,23 @@
 import {block} from '../../../utils/cn';
-import type {StoriesItemMedia} from '../../types';
+import type {MediaRendererProps} from '../MediaRenderer/MediaRenderer';
 
 import './ImageView.scss';
 
 const b = block('stories-image-view');
 
-export interface ImageViewProps {
-    media: StoriesItemMedia;
-}
+export interface ImageViewProps extends MediaRendererProps {}
 
-export function ImageView({media}: ImageViewProps) {
-    const type = media.type || 'image';
+export function ImageView({media, style}: ImageViewProps) {
+    if (media.type === 'image') {
+        return (
+            <div
+                className={b({style})}
+                style={{
+                    background: `url("${media.url2x ?? media.url}") left 0 top 0 / cover no-repeat`,
+                }}
+            />
+        );
+    }
 
-    return type === 'image' ? <img className={b()} src={media.url} alt="" /> : null;
+    return null;
 }
