+++ src/components/Stories/components/VideoView/VideoView.scss
@@ -3,8 +3,41 @@
 $block: '.#{variables.$ns}stories-video-view';
 
 #{$block} {
-    width: auto;
+    width: 100%;
+    height: 100%;
     max-width: 100%;
     max-height: 100%;
-    margin: auto;
+
+    position: relative;
+    overflow: hidden;
+
+    &_style_half-size {
+        border-radius: var(--g-spacing-5);
+    }
+
+    &_style_half-size-with-margins {
+        width: calc(100% - 2 * var(--g-spacing-2));
+        height: calc(100% - 2 * var(--g-spacing-2));
+
+        border-radius: var(--g-spacing-5);
+        margin: var(--g-spacing-2);
+    }
+
+    &_style_full-size {
+        position: absolute;
+        inset-block-start: 0;
+        inset-inline-start: 0;
+        border-radius: var(--g-spacing-5);
+
+        z-index: 0;
+    }
+
+    &__video {
+        position: absolute;
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+
+        z-index: 0;
+    }
 }
