+++ src/components/Stories/types.ts
@@ -1,9 +1,11 @@
 import type * as React from 'react';
+import {ButtonProps} from '@gravity-ui/uikit';
 
 export type StoriesItemMedia = {url: string} & (
     | {
-          /** default 'image' */
-          type?: 'image';
+          type: 'image';
+          /** url for img srcSet props - apply on Retina display */
+          url2x?: string;
       }
     | {
           type: 'video';
@@ -12,11 +14,31 @@ export type StoriesItemMedia = {url: string} & (
       }
 );
 
+export enum StoriesTextBlockStyle {
+    Card = 'card',
+    Transparent = 'transparent',
+}
+
+export enum StoriesMediaBlockStyle {
+    HalfSizeWithMargins = 'half-size-with-margins',
+    HalfSize = 'half-size',
+    FullSize = 'full-size',
+}
+
+export interface StoriesItemTextColorStyles {
+    titleColor?: string;
+    descriptionColor?: string;
+    counterColor?: string;
+}
+
 export interface StoriesItem {
     title?: string;
-    /** @deprecated use `content` property instead */
     description?: string;
-    content?: React.ReactNode;
+    textColorStyles?: StoriesItemTextColorStyles;
+    textBlockStyle: StoriesTextBlockStyle;
+    mediaBlockStyle: StoriesMediaBlockStyle;
+    firstAction?: ButtonProps;
+    secondAction?: ButtonProps;
     /** Url for link "more" */
     url?: string;
     media?: StoriesItemMedia;
