+++ src/components/StoriesGroup/StoriesGroup.scss
@@ -7,6 +7,10 @@ $borderRadius: 20px;
     --g-modal-border-radius: #{$borderRadius};
     --g-modal-margin: 20px;
 
+    & .g-modal__content-wrapper {
+        overflow-x: initial;
+    }
+
     &__modal-content {
         background-color: inherit;
         border-radius: $borderRadius;
