--- src/components/unstable/Stories/components/ImageView/ImageView.scss
@@ -1,33 +1,0 @@
-@use '../../../variables';
-
-$block: '.#{variables.$ns}stories-image-view';
-
-$borderRadius: 20px;
-
-#{$block} {
-    width: 100%;
-    height: 100%;
-    max-width: 100%;
-    max-height: 100%;
-
-    &_style_half-size {
-        border-radius: $borderRadius;
-    }
-
-    &_style_half-size-with-margins {
-        width: calc(100% - 2 * var(--g-spacing-2));
-        height: calc(100% - 2 * var(--g-spacing-2));
-        border-radius: $borderRadius;
-
-        margin: var(--g-spacing-2);
-    }
-
-    &_style_full-size {
-        position: absolute;
-        border-radius: $borderRadius;
-
-        inset-block-start: 0;
-        inset-inline-start: 0;
-        z-index: 0;
-    }
-}
