--- src/components/unstable/Stories/components/VideoView/VideoView.scss
@@ -1,43 +1,0 @@
-@use '../../../variables';
-
-$block: '.#{variables.$ns}stories-video-view';
-
-#{$block} {
-    width: 100%;
-    height: 100%;
-    max-width: 100%;
-    max-height: 100%;
-
-    position: relative;
-    overflow: hidden;
-
-    &_style_half-size {
-        border-radius: var(--g-spacing-5);
-    }
-
-    &_style_half-size-with-margins {
-        width: calc(100% - 2 * var(--g-spacing-2));
-        height: calc(100% - 2 * var(--g-spacing-2));
-
-        border-radius: var(--g-spacing-5);
-        margin: var(--g-spacing-2);
-    }
-
-    &_style_full-size {
-        position: absolute;
-        inset-block-start: 0;
-        inset-inline-start: 0;
-        border-radius: var(--g-spacing-5);
-
-        z-index: 0;
-    }
-
-    &__video {
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        object-fit: cover;
-
-        z-index: 0;
-    }
-}
