--- src/components/unstable/Stories/components/index.ts
@@ -1,3 +1,0 @@
-export * from './ImageView/ImageView';
-export * from './VideoView/VideoView';
-export * from './MediaRenderer/MediaRenderer';
