--- src/components/unstable/Stories/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useSyncWithLS';
