--- src/components/unstable/Stories/hooks/useSyncWithLS/index.ts
@@ -1,2 +1,0 @@
-export {useSyncWithLS} from './useSyncWithLS';
-export type {UseSyncWithLSInputProps, UseSyncWithLSOutputProps} from './useSyncWithLS';
