--- src/components/unstable/Stories/i18n/en.json
@@ -1,6 +1,0 @@
-{
-  "label_back": "Back",
-  "label_next": "Next",
-  "label_close": "Close",
-  "label_more": "More"
-}
