--- src/components/unstable/Stories/i18n/ru.json
@@ -1,6 +1,0 @@
-{
-  "label_back": "Назад",
-  "label_next": "Дальше",
-  "label_close": "Закрыть",
-  "label_more": "Подробнее"
-}
