--- src/components/unstable/Stories/index.ts
@@ -1,3 +1,0 @@
-export * from './Stories';
-export * from './hooks';
-export * from './types';
