--- src/components/unstable/StoriesGroup/components/StoriesPreview/StoriesPreview.scss
@@ -1,65 +1,0 @@
-@use '../../../variables';
-
-$block: '.#{variables.$ns}stories-group-preview';
-
-#{$block} {
-    user-select: none;
-    padding-block-end: 16px;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 8px;
-
-    &__slider-preview-list-wrapper {
-        display: flex;
-        gap: 8px;
-    }
-
-    &__stories-preview-item {
-        $disabled: #{&}_disabled;
-
-        position: relative;
-        box-sizing: border-box;
-
-        display: flex;
-        align-items: center;
-        justify-content: center;
-
-        cursor: pointer;
-
-        overflow: hidden;
-        width: 40px;
-        height: 40px;
-        border-radius: 8px;
-
-        background-color: var(--g-color-base-background);
-
-        &::after {
-            content: '';
-            position: absolute;
-            inset-block-start: 0;
-            inset-inline-start: 0;
-
-            background-color: var(--g-color-sfx-shadow);
-            width: 40px;
-            height: 40px;
-        }
-
-        &_active,
-        &:hover:not(#{$disabled}) {
-            &::after {
-                display: none;
-            }
-        }
-
-        &_active {
-            border: 2px solid var(--g-color-line-brand);
-        }
-    }
-
-    &__slider-button-wrapper {
-        width: 28px;
-        height: 28px;
-    }
-}
