--- src/components/unstable/StoriesGroup/components/index.ts
@@ -1,1 +1,0 @@
-export * from './StoriesPreview/StoriesPreview';
