--- src/components/unstable/StoriesGroup/index.ts
@@ -1,2 +1,0 @@
-export * from './StoriesGroup';
-export * from './types';
