--- src/components/unstable/index.ts
@@ -1,17 +1,0 @@
-/* eslint-disable camelcase */
-export {
-    type StoriesItemMedia as unstable_StoriesItemMedia,
-    type StoriesItem as unstable_StoriesItem,
-    type StoriesProps as unstable_StoriesProps,
-    Stories as unstable_Stories,
-    useSyncWithLS as unstable_useSyncWithLS,
-    StoriesMediaBlockStyle,
-    StoriesTextBlockStyle,
-    StoriesItemTextColorStyles,
-} from './Stories';
-
-export {
-    type StoriesGroupItem,
-    type StoriesGroupProps,
-    StoriesGroup as unstable_StoriesGroup,
-} from './StoriesGroup';
