+++ src/components/utils/cn.ts
@@ -3,7 +3,6 @@ import {withNaming} from '@bem-react/classname';
 export type CnMods = Record<string, string | boolean | undefined>;
 
 export const NAMESPACE = 'gc-';
-export const UNSTABLE_NAMESPACE = 'gcu-';
 
 export const cn = withNaming({e: '__', m: '_', v: '_'});
 export const block = withNaming({n: NAMESPACE, e: '__', m: '_', v: '_'});
